# graphite: reduce the memory used by GraphiteParser.Parse() by ~80%