@charlievieth 2019-05-13
Will investigate going line by line and will update the PR if it improves perf or note why it didn't (and will remove my comment if thats the case).