@charlievieth 2019-05-13
@glinton updated to step through the buffer line-by-line the improves perf by roughly ~5%.